backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
Pillow>=10.3.0
openai>=1.12.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
